# api/v1/client/swagger/api/swagger.yaml
openapi: 3.0.0
info:
  title: API v1
  description: "# Introduction\nWelcome to the documentation for Hatchet's API. \n"
  termsOfService: https://docs.hatchet.run/terms-and-privacy/terms-of-service/
  contact:
    name: Hatchet Support
    email: support@hatchet.run
  version: 1.0.0
servers:
- url: http://localhost:8080
security:
- Bearer: []
...
paths:
...
  /api/v1/teams/{team_id}/modules:
    get:
      tags:
      - Modules
      summary: List Modules
      description: Lists modules for a given team.
      operationId: listModules
      parameters:
      - name: page
        in: query
        description: The page to query for
        required: false
        style: form
        explode: true
        schema:
          type: integer
          format: int64
          x-go-name: Page
        x-go-name: Page
      - name: team_id
        in: path
        description: The team id
        required: true
        style: simple
        explode: false
        schema:
          type: string
          x-go-name: Team
        x-go-name: Team
      responses:
        "200":
          description: Successfully listed modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListModulesResponse'
        "400":
          description: A malformed or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorBadRequestExample'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorForbiddenExample'
      x-codeSamples:
      - lang: Shell
        source: |-
          curl --request GET \
            --url 'http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules?page=2' \
            --header 'Authorization: Bearer REPLACE_BEARER_TOKEN'
      - lang: Node
        source: |-
          const fetch = require('node-fetch');

          let url = 'http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules?page=2';

          let options = {method: 'GET', headers: {Authorization: 'Bearer REPLACE_BEARER_TOKEN'}};

          fetch(url, options)
            .then(res => res.json())
            .then(json => console.log(json))
            .catch(err => console.error('error:' + err));
      - lang: Ruby
        source: |-
          require 'uri'
          require 'net/http'

          url = URI("http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules?page=2")

          http = Net::HTTP.new(url.host, url.port)

          request = Net::HTTP::Get.new(url)
          request["Authorization"] = 'Bearer REPLACE_BEARER_TOKEN'

          response = http.request(request)
          puts response.read_body
      - lang: Python
        source: |-
          import requests

          url = "http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules"

          querystring = {"page":"2"}

          headers = {"Authorization": "Bearer REPLACE_BEARER_TOKEN"}

          response = requests.request("GET", url, headers=headers, params=querystring)

          print(response.text)
      - lang: Go
        source: "package main\n\nimport (\n\t\"fmt\"\n\t\"net/http\"\n\t\"io/ioutil\"\
          \n)\n\nfunc main() {\n\n\turl := \"http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules?page=2\"\
          \n\n\treq, _ := http.NewRequest(\"GET\", url, nil)\n\n\treq.Header.Add(\"\
          Authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n\n\tres, _ := http.DefaultClient.Do(req)\n\
          \n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\n\t\
          fmt.Println(res)\n\tfmt.Println(string(body))\n\n}"
    post:
      tags:
      - Modules
      summary: Create Module
      description: Creates a new module.
      operationId: createModule
      parameters:
      - name: team_id
        in: path
        description: The team id
        required: true
        style: simple
        explode: false
        schema:
          type: string
          x-go-name: Team
        x-go-name: Team
      requestBody:
        description: The module to create
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/createModule_request'
        required: false
      responses:
        "201":
          description: Successfully created the module
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateModuleResponse'
        "400":
          description: A malformed or bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorBadRequestExample'
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIErrorForbiddenExample'
      x-codeSamples:
      - lang: Shell
        source: |-
          curl --request POST \
            --url http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules \
            --header 'Authorization: Bearer REPLACE_BEARER_TOKEN' \
            --header 'content-type: application/json' \
            --data '{"github":{"github_app_installation_id":"bb214807-246e-43a5-a25d-41761d1cff9e","github_repository_branch":"main","github_repository_name":"infra","github_repository_owner":"hatchet-dev","path":"./staging/eks"},"name":"string"}'
      - lang: Node
        source: |-
          const fetch = require('node-fetch');

          let url = 'http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules';

          let options = {
            method: 'POST',
            headers: {
              Authorization: 'Bearer REPLACE_BEARER_TOKEN',
              'content-type': 'application/json'
            },
            body: '{"github":{"github_app_installation_id":"bb214807-246e-43a5-a25d-41761d1cff9e","github_repository_branch":"main","github_repository_name":"infra","github_repository_owner":"hatchet-dev","path":"./staging/eks"},"name":"string"}'
          };

          fetch(url, options)
            .then(res => res.json())
            .then(json => console.log(json))
            .catch(err => console.error('error:' + err));
      - lang: Ruby
        source: |-
          require 'uri'
          require 'net/http'

          url = URI("http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules")

          http = Net::HTTP.new(url.host, url.port)

          request = Net::HTTP::Post.new(url)
          request["Authorization"] = 'Bearer REPLACE_BEARER_TOKEN'
          request["content-type"] = 'application/json'
          request.body = "{\"github\":{\"github_app_installation_id\":\"bb214807-246e-43a5-a25d-41761d1cff9e\",\"github_repository_branch\":\"main\",\"github_repository_name\":\"infra\",\"github_repository_owner\":\"hatchet-dev\",\"path\":\"./staging/eks\"},\"name\":\"string\"}"

          response = http.request(request)
          puts response.read_body
      - lang: Python
        source: |-
          import requests

          url = "http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules"

          payload = {
              "github": {
                  "github_app_installation_id": "bb214807-246e-43a5-a25d-41761d1cff9e",
                  "github_repository_branch": "main",
                  "github_repository_name": "infra",
                  "github_repository_owner": "hatchet-dev",
                  "path": "./staging/eks"
              },
              "name": "string"
          }
          headers = {
              "Authorization": "Bearer REPLACE_BEARER_TOKEN",
              "content-type": "application/json"
          }

          response = requests.request("POST", url, json=payload, headers=headers)

          print(response.text)
      - lang: Go
        source: "package main\n\nimport (\n\t\"fmt\"\n\t\"strings\"\n\t\"net/http\"\
          \n\t\"io/ioutil\"\n)\n\nfunc main() {\n\n\turl := \"http://localhost:8080/api/v1/teams/322346f9-54b4-497d-bc9a-c54b5aaa4400/modules\"\
          \n\n\tpayload := strings.NewReader(\"{\\\"github\\\":{\\\"github_app_installation_id\\\
          \":\\\"bb214807-246e-43a5-a25d-41761d1cff9e\\\",\\\"github_repository_branch\\\
          \":\\\"main\\\",\\\"github_repository_name\\\":\\\"infra\\\",\\\"github_repository_owner\\\
          \":\\\"hatchet-dev\\\",\\\"path\\\":\\\"./staging/eks\\\"},\\\"name\\\"\
          :\\\"string\\\"}\")\n\n\treq, _ := http.NewRequest(\"POST\", url, payload)\n\
          \n\treq.Header.Add(\"Authorization\", \"Bearer REPLACE_BEARER_TOKEN\")\n\
          \treq.Header.Add(\"content-type\", \"application/json\")\n\n\tres, _ :=\
          \ http.DefaultClient.Do(req)\n\n\tdefer res.Body.Close()\n\tbody, _ := ioutil.ReadAll(res.Body)\n\
          \n\tfmt.Println(res)\n\tfmt.Println(string(body))\n\n}"
      x-codegen-request-body-name: CreateModuleRequest
...
components:
  schemas:
...
    APIErrorBadRequestExample:
      type: object
      properties:
        code:
          type: integer
          description: a custom Hatchet error code
          format: uint64
          example: 1400
          x-go-name: Code
        description:
          type: string
          description: a description for this error
          example: Bad request (detailed error)
          x-go-name: Description
        docs_link:
          type: string
          description: "a link to the documentation for this error, if it exists"
          example: github.com/hatchet-dev/hatchet
          x-go-name: DocsLink
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
    APIErrorForbiddenExample:
      type: object
      properties:
        code:
          type: integer
          description: a custom Hatchet error code
          format: uint64
          example: 1403
          x-go-name: Code
        description:
          type: string
          description: a description for this error
          example: Forbidden
          x-go-name: Description
        docs_link:
          type: string
          description: "a link to the documentation for this error, if it exists"
          example: github.com/hatchet-dev/hatchet
          x-go-name: DocsLink
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
...
    CreateModuleRequestGithub:
      required:
      - github_app_installation_id
      - github_repository_branch
      - github_repository_name
      - github_repository_owner
      - path
      type: object
      properties:
        github_app_installation_id:
          type: string
          description: "this refers to the Hatchet app installation id, **not** the\
            \ installation id stored on Github"
          example: bb214807-246e-43a5-a25d-41761d1cff9e
          x-go-name: GithubAppInstallationID
        github_repository_branch:
          type: string
          description: the repository branch on Github
          example: main
          x-go-name: GithubRepositoryBranch
        github_repository_name:
          type: string
          description: the repository name on Github
          example: infra
          x-go-name: GithubRepositoryName
        github_repository_owner:
          type: string
          description: the repository owner on Github
          example: hatchet-dev
          x-go-name: GithubRepositoryOwner
        path:
          type: string
          description: path to the module in the github repository
          example: ./staging/eks
          x-go-name: Path
      example:
        github_repository_branch: main
        path: ./staging/eks
        github_app_installation_id: bb214807-246e-43a5-a25d-41761d1cff9e
        github_repository_owner: hatchet-dev
        github_repository_name: infra
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
    CreateModuleResponse:
      $ref: '#/components/schemas/Module'
...
    ListModulesResponse:
      type: object
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationResponse'
        rows:
          type: array
          items:
            $ref: '#/components/schemas/Module'
          x-go-name: Rows
      example:
        pagination:
          next_page: 3
          num_pages: 10
          current_page: 2
        rows:
        - updated_at: 2022-12-13T15:06:48.888358-05:00
          name: eks
          created_at: 2022-12-13T15:06:48.888358-05:00
          id: bb214807-246e-43a5-a25d-41761d1cff9e
          deployment:
            path: path
            github_app_installation_id: github_app_installation_id
            github_repo_name: github_repo_name
            github_repo_branch: github_repo_branch
            github_repo_owner: github_repo_owner
        - updated_at: 2022-12-13T15:06:48.888358-05:00
          name: eks
          created_at: 2022-12-13T15:06:48.888358-05:00
          id: bb214807-246e-43a5-a25d-41761d1cff9e
          deployment:
            path: path
            github_app_installation_id: github_app_installation_id
            github_repo_name: github_repo_name
            github_repo_branch: github_repo_branch
            github_repo_owner: github_repo_owner
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
...
    Module:
      type: object
      properties:
        created_at:
          type: string
          description: the time that this resource was created
          format: date-time
          example: 2022-12-13T15:06:48.888358-05:00
          x-go-name: CreatedAt
        deployment:
          $ref: '#/components/schemas/ModuleDeploymentConfig'
        id:
          type: string
          description: "the id of this resource, in UUID format"
          example: bb214807-246e-43a5-a25d-41761d1cff9e
          x-go-name: ID
        name:
          type: string
          description: the name for the module
          example: eks
          x-go-name: Name
        updated_at:
          type: string
          description: the time that this resource was last updated
          format: date-time
          example: 2022-12-13T15:06:48.888358-05:00
          x-go-name: UpdatedAt
      example:
        updated_at: 2022-12-13T15:06:48.888358-05:00
        name: eks
        created_at: 2022-12-13T15:06:48.888358-05:00
        id: bb214807-246e-43a5-a25d-41761d1cff9e
        deployment:
          path: path
          github_app_installation_id: github_app_installation_id
          github_repo_name: github_repo_name
          github_repo_branch: github_repo_branch
          github_repo_owner: github_repo_owner
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
    ModuleDeploymentConfig:
      type: object
      properties:
        github_app_installation_id:
          type: string
          x-go-name: GithubAppInstallationID
        github_repo_branch:
          type: string
          x-go-name: GithubRepoBranch
        github_repo_name:
          type: string
          x-go-name: GithubRepoName
        github_repo_owner:
          type: string
          x-go-name: GithubRepoOwner
        path:
          type: string
          x-go-name: Path
      example:
        path: path
        github_app_installation_id: github_app_installation_id
        github_repo_name: github_repo_name
        github_repo_branch: github_repo_branch
        github_repo_owner: github_repo_owner
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
...
    PaginationResponse:
      type: object
      properties:
        current_page:
          type: integer
          description: the current page
          format: int64
          example: 2
          x-go-name: CurrentPage
        next_page:
          type: integer
          description: the next page
          format: int64
          example: 3
          x-go-name: NextPage
        num_pages:
          type: integer
          description: the total number of pages for listing
          format: int64
          example: 10
          x-go-name: NumPages
      example:
        next_page: 3
        num_pages: 10
        current_page: 2
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
...
    createModule_request:
      type: object
      properties:
        github:
          $ref: '#/components/schemas/CreateModuleRequestGithub'
        name:
          type: string
          x-go-name: Name
      x-go-package: github.com/hatchet-dev/hatchet/api/v1/types
...
  securitySchemes:
    Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
...